// src/types/schemas/currencyConversionRate.d.ts
/* prettier-ignore */
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type CurrencyConversionRate = null | {
	rate: number;
	validUntil: string;
	estimatedRefreshTime: string;
	seriesKey: string;
};